+++ jest.config.js
@@ -1,6 +1,0 @@
-/* eslint-env node */
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  collectCoverageFrom: ['src/**/*.ts'],
-};