+++ src/browseraction/browseraction.ts
@@ -1,0 +1,6 @@
+import './assets/browseraction.css'
+
+import { createApp } from 'vue'
+import App from './BrowserAction.vue'
+
+createApp(App).mount('#app')