+++ src/browseraction/components/store/browser-storage.ts
@@ -1,0 +1,8 @@
+export enum BrowserStorageKey {
+  urlList = 'txt',
+  lazyload = 'lazyload',
+  random = 'random',
+  reverse = 'reverse',
+  preserve = 'preserve',
+  deduplicate = 'deduplicate'
+}