+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.0
+
+- Added
+  - Option to ignore duplicate URLs on open
+- Changed
+  - Complete rewrite in Vue.js
+
 ## 1.5.6
 
 - Fixed